Moved custom themed stories to custom themed folder and changed chromatic so snapshots for custom themed stories are taken with the base theme
