+++ src/js/components/Card/stories/CustomThemed/RichFooter.js
@@ -102,5 +102,5 @@ export const RichFooter = () => {
 RichFooter.storyName = 'Rich footer';
 
 export default {
-  title: `Layout/Card/Rich footer`,
+  title: `Layout/Card/Custom Themed/Rich footer`,
 };
