+++ src/js/components/RadioButtonGroup/stories/CustomThemed/CustomTheme.js
@@ -67,5 +67,5 @@ export const CustomRadioButtonGroup = ({ value: initialValue, ...props }) => {
 CustomRadioButtonGroup.storyName = 'Custom theme';
 
 export default {
-  title: 'Input/RadioButtonGroup/Custom themed/Custom theme',
+  title: 'Input/RadioButtonGroup/Custom Themed/Custom theme',
 };
